requirements.txt
discord.py
asyncpg
jishaku
sr_api
asyncpraw
waifuim.py
psutil
asyncio
python-dotenv
